architecture: introduce ->parse_command_line()
Introduce the pure virtual architecture::parse_command_line() for parsing
compiler command lines.

Actual class architecture implementations represent some compiler in a
specific configuration and are the place to parse and evaluate original
compiler invocations.

In general, the set of builtin macros depends on the outcome.
Drop architecture::register_builtins() in favor of this new
architecture::parse_command_line() -- among other things, it is supposed to
register the builtin macros with the given preprocessor instance as
appropriate.

The testcases are made to invoke it with a nullptr argv argument to request
a default set suitable for testing.

Finally, make class arch_x86_64_gcc provide an implementation of
architecture::parse_command_line() specific to GCC.

Signed-off-by: Nicolai Stange <nstange@suse.de>