gnuc_parser: fix location computation for productions beginning with optionals
There are productions involving optional subproductions all over the
grammar. If not matching anything, these usually invoke the empty()
function which sets their locations' ->begin to ->end. Because
YYLLOC_DEFAULT() initializes the locations for empties with the range of
the previously matched (sub-)productions, this achieves the desired
result if the empty sub-production appears anywhere but at the beginning
of its parent production.

Explicitly fixup the locations for productions beginning with possibly
empty subproductions, if necessary.

Signed-off-by: Nicolai Stange <nstange@suse.de>