gcc_cmdline_parser: always search for non-inverted option variants first
Currently, the gcc_cmdline_parser always interprets options of the form
-fno-foo, -mno-foo or -Wno-foo as specifying the inverted variants of
-ffoo, -mfoo or -Wfoo respectively.

However, for e.g. the case -mno-red-zone, -mred-zone is not a valid option
and thus, the gcc_cmdline_parser would report an invalid option.

Fix this by always searching for the full option name, including the "no-"
prefix, if any, first and fall back to the inverted option handling only
if that fails.

Signed-off-by: Nicolai Stange <nstange@suse.de>