preprocessor: record a conditional_inclusion_node's taken branch, if any
A conditional_inclusion_node instance already stores the
raw_pp_tokens_range starting from the opening #if/#ifdef/#ifndef and
spanning up to the closing #endif. Also, a list of all involved directive
ranges is kept.

Recording the index of the taken branch, if any, will enable the
determination of the raw_pp_tokens subrange associated with that branch
and, of equal importance, this subrange's complement.

Introduce the new, 1-based ->_taken_branch index member to class
conditional_inclusion_node. It will be set to zero if none of the
branches had been taken.

Make the preprocessor maintain ->_taken_branch.

Signed-off-by: Nicolai Stange <nstange@suse.de>