depreprocessor: emit extra catch-all #else directives as a safety measure
Currently, the depreprocessor amends an #error directive for non-taken
conditional branches retained from the input.

However, in case that there is some branch taken, but no #else branch
presnt in the input code, no #error will be hit in the unexpected case
that none of the branches are taken at all when compiling the output.

As a safety measure, make the depreprocessor emit an extra #else branch
containing an #error directive for these cases.

Signed-off-by: Nicolai Stange <nstange@suse.de>