create_lp: include headers containing only preprocessing directives
Currently, headers are included from the final live patch only if they
provide some needed AST entity, but not for needed #defines or #undefs.
The reason is that the depreprocessing happens after the AST dependency
resolving phase and that new #includes can introduce additional ones.

However, this is obviously not the case for headers containing nothing
but preprocessing directives. Always include these (if eligible) in the
hope that they provide some macro definitions needed by the final live
patch.

Signed-off-by: Nicolai Stange <nstange@suse.de>