preprocessor: track full directive range for #endifs
Currently, the preprocessor records a raw_pp_tokens_range terminating right
after the "endif" identifier for #endif directives.

The reason is that for the particular case of #endifs, it doesn't consume
input until the the terminating newline in order to avoid reading an EOF
and triggering the associated error reporting due to aconditional inclusion
being active. However, this causes inconsistencies with the other
conditional inclusion related directives for which the terminating newline
is always covered by the recorded raw_pp_tokens_range.

Improve the preprocessor's #endif handling code to always include the
terminating newline in their associated raw_pp_tokens_range while being
careful not to trigger any error reporting due to EOFs.

Signed-off-by: Nicolai Stange <nstange@suse.de>