preprocessor: simplify used macro handling for the defined() operator
The preprocessor goes to great lengths to maintain macro_invocation
instances for the defined() operator itself as well as its arguments.

However, it's a bit of an inconsistency to have a macro_invocation
for the argument to the defined() operator, because it won't ever get
expanded.

Furthermore, all the work is in vain anyways: the defined() operator
can only be part of a preprocessing conditional and any macro_invocation
instantiated in the course of processing these will get dropped when
pp_result::_drop_pp_tokens_tail() gets called eventually.

Simplify the defined() operator handling by directly recording the
argument either as a used_macro or a macro_nondef_constraint at the
corresponding conditional_inclusion_node rather than at some
macro_invocation.

Signed-off-by: Nicolai Stange <nstange@suse.de>