depreprocessor: implement support for macro invocations' argument rewriting
Currently, the depreprocessor lacks any macro invocation rewrite support
whatsoever. In particular, it would not retain any macro invocation where
some of the tokens from its expansion are to be replaced in the resulting
livepatch, but emit the expanded sequence instead.

However, in case the to be replaced token corresponds to some of the
macro's arguments which has been passed through without modifications
like concatenation or stringifcation, support for replacing argument tokens
would allow for the macro invocation to be kept. Note that this is the most
relevant case, as identifiers subject to renaming tend to be given
arguments to e.g. READ_ONCE() from the Linux kernel rather than being
hardcoded directly in some macro definition.

Implement argument rewriting for macro invocations.

The preprocessor code has got support for tracking passed through macro
argument tokens with the previous commits already.

Make the depreprocessor search for sequences of consecutive action::copy
and action::replace transformed_input_chunk::_op instances overlapping with
some macro_invocation for which all of the action::replace operations can
be satisfied by (consistent) macro invocation argument rewrites.

The subrange of a found such sequence which stems from the macro
invocation is then replaced by an instance of the new
action::rewrite_macro_invocation _op for describing the macro argument
token replacements to be made in the final output.

Signed-off-by: Nicolai Stange <nstange@suse.de>