target_x86_64_gcc: clarify comment in evaluate_enum_type()
A comment preceeding the enum member type evaluation code in
target_x86_64_gcc::evaluate_enum_type() states that GCC's behaviour is
"a bit strange", when it in fact isn't: the rule is to conform to the
C standard and choose an int whenever possible, or, as an extension,
the (wider) enum's underlying integer type otherwise. C.f.
GCC's finish_enum(). Clarify the comment.

Signed-off-by: Nicolai Stange <nstange@suse.de>