user_policy_command: ignore EINTR from select(2)
Currently, if the select(2) call from user_policy_command::instance::wait()
returns any kind of error, processing of the child's output is aborted and
an exception thrown. For EINTR caused by e.g. a SIGCHLD, this behaviour
is wrong. As default signal actions are not overridden, said select(2)
returning early with EINTR should simply be retried. Implement this.

Signed-off-by: Nicolai Stange <nstange@suse.de>