target: move float_mode_to_float_kind() to target_gcc
That floating modes from float_mode_kind are actually an implementation
detail of GCC and should be moved out of the generic, compiler agnostic
code. As a preparation for that, move the pure virtual
float_mode_to_float_kind() to class target_gcc. As only that class'
implementation itself make use of this function, make it also private.

Signed-off-by: Nicolai Stange <nstange@suse.de>