types: make complex_float_type inherit directly from arithmetic_type
The common ancestor class float_type of real_float_type and
complex_float_type is almost unused and the common ancestry between the
two doesn't bring any benefits. In working towards eliminating it, make
class complex_float_type a direct descendant of class arithmetic_type.

In order to get rid of the dependency on the float_type::kind enum,
reimplement it to refer to its' component's underlying type by means
of a std::shared_ptr<const real_float_type> in place of the former
float_type::kind which used to be stored within the base float_type
instance. As a side-effect, this change will also lay the grounds for more
generic complex type where the components are not of any of the standard
float types, like e.g. the extended float types or even integer types as
supported by GCC.

Signed-off-by: Nicolai Stange <nstange@suse.de>