types: double-dispatch arithmetic conversions only down to int_type
arithmetic::arithmetic_conversion() implements a double-dispatch pattern
for routing down to the actual conversion implementations specific to pairs
of leaf types. Currently, it is overloaded on std_int_type and defaults to
promoting the rest, i.e. plain_char_type, bool_type and enum_type.

However, in lieu of a richer hierarchy below int_type for representing the
extended integer types, it is desirable to delegate all conversion between
integer types to int_type::intege_conversion(), which implements a similar
double-dispatch scheme, but restricted to the integer types.

Replace the overloads of arithmetic_type::arithmetic_conversion() on
std_int_type with overloads on the int_type base class and delegate
conversions between integers to the int_type::integer_conversion()
double-dispatch.

Signed-off-by: Nicolai Stange <nstange@suse.de>