ast_evaluate: use int_type for results of int_type::promote()
Currently, int_type::promote() returns a std_int_type and the AST evalutor
stores results therefrom in pointers/references of that type at a couple of
place. In order to prepare for int_type::promote() being made to return an
abstract int_type, make the AST evaluators to use that type for the
obtained therefrom.

Signed-off-by: Nicolai Stange <nstange@suse.de>