preprocessor: maintain order for inserted whitespace tokens
The sequence of pp_tokens is supposed to be sorted with respect to
their raw_pp_tokens_range. However, when inserting extra whitespace between
certain punctuators for the sake of preprocessing idempotency,
preprocessor::::read_next_token() simply assigns an empty range with
begin and end set to the subsequent token's begin. If the previous token
happens to stem from the very same macro invocation as the subsequent one,
this will break the sortedness invariant: the previous token compares equal
to the subsequent one, but the whitespace inserted in the middle compares
as less than the next one. As a consequence, the binary search in
pp_result::raw_pp_tokens_range_to_nonraw() can get confused and yield wrong
results.

Fix this by assigning a correct raw_pp_tokens_range to the inserted
whitespace token in case the previous and next tokens stem from the same
macro invocation.

Signed-off-by: Nicolai Stange <nstange@suse.de>