create_lp: introduce support for basic dead code elimination
There are various C language constructs where the the unreachability of
child expressions can get trivially recognized based on constant expression
evaluation, for example the ternary operator or __builtin_choose_expr() to
name only a few.

GCC also implements this at any (or low) optimization levels and
subsequently purges all references to potentially undefined functions and
objects from dead code. If klp-ccp attempts to add such an undefined
function or object to the closure, it will inevitably fail to do so.

Instead, it should only add dependencies to the declarations for references
from expressions known to never get executed during execution such that the
resulting livepatch still compiles. Note that this is similar to how sizeof
or __alignof__ expressions are handled nowadays already.

Introduce a simple facility for tracking unevaluated, i.e. never executed
code during create_lp's AST info collection traversal. For ease of
implementation, it is maintained in the form of a stack of (pointers to)
AST nodes, where any AST node on the stack is considered to be never
evaluated during execution

Reimplement the current sizeof()/__alignof__() handling on the grounds of
this.

Signed-off-by: Nicolai Stange <nstange@suse.de>