AST: keep track of whether a struct/union member is the last one
The last member in a struct may be an array of unspecified length. Currently,
all of a struct's members are getting processed in one go in
struct_declaration_list::create_content() and this can determine whether some
member is the last one or not per its complete view of the struct contents.

However, member handling will get integrated into the evaluation code's main
AST walk in a future commit and struct_declaration_list::create_content() get
removed. Introduce a means for tracking whether or not some struct_declarator
member comes last so that the AST walking code can recognize it by inspection
of the member already.

Signed-off-by: Nicolai Stange <nstange@suse.de>