create_lp: fix crash on function defintition references from invalid contexts
Even though invalid C, the input code could still reference a function
definition in invalid contexts as e.g. an array length specifier at file scope.

The AST info collection code in create_lp() would crash when registering a
dependency to a function definition in such a case, because there is no
depedency tracking object available, that is, _cur_deps is NULL. Note that
there are already checks in place for the case a reference resolves to some
object or function init_declarator, but function definitions have been missed.
Fix this.

Signed-off-by: Nicolai Stange <nstange@suse.de>