target_gcc: rename sou_layouter's internal variables to clarify units
The implementation of target_gcc::sou_layouter maintains several variables
corrsponding to offsets or alignments resp.. These are either given in
units of bytes or bits and it is not always immediately obvious at first
sight what the correct one is. Rename a couple of
target_gcc::sou_layouter's internal variables to include a clear indication
on the unit.

Signed-off-by: Nicolai Stange <nstange@suse.de>