cpp: allow for specification of compiler to emulate + its invocation
Currently, the cpp tool is used only from the testsuite and the compiler
to emulate is hardcoded as "x86_64-gcc-4.8.5". Moreover, it is not
possible to specify the compiler invocation.

In order to make the cpp tool more useful for debugging, extend its
command line to allow for the specification of the compiler to emulate as
well as its invocation, just in analogy to the klp-ccp tool.

Signed-off-by: Nicolai Stange <nstange@suse.de>