(de)preprocessor: support conditional inclusions in macro invocations
All of the (de)preprocessor code is based on the premise that a single
macro invocation cannot interleave with conditional inclusion directives.

This however is wrong, at least as far as GCC is concerned. While
C99 6.10.3 ("Macro replacement") explicitly states that preprocessor
directives within the list of a func-like macro invocation's argument
result in undefined behaviour, GCC still supports those in the case
if conditional inclusions (#if, #ifdef, #else, #endif) with the semantics
that the conditional inclusions get evaluated before any macro expansion
takes place.

That is,

  #define A(x) x
  #if 1
  A(
  #endif
    #if 1
      1
    #else
      0
    #endif
  #if
  )
  #endif

is logically equivalent to A(1).

The preprocessor code is almost ready to support conditional inclusion
directives interleaving with macro invocations, except for that if any of
the directives' conditional expressions in turn invokes some macro, it will
get confused and crash. That's because the preprocessor's internal
->_cur_macro_invocation has been set to a fresh macro_invocation instance
for the outer macro invocation before its argument list gets peeked for
in the (raw) token stream. The inner macro invocation will hijack that
macro_invocation instance and, once its expansion has finished, wrongly
reset ->_cur_macro_invocation to nullptr. This can be easily fixed by
saving ->_cur_macro_invocation away before processing conditional inclusion
directives and restoring it afterwards again.

For the depreprocessor, the needed changes are much more involved though.
A conditional inclusion directive within some transformed_input_chunk's
spanned range is described by a transformed_input_chunk::_op instance
of type action::cond_incl_transition. Note that a transformed_input_chunk's
sequence of _op instances must be sorted by both, the associated
pp_tokens_range as well as the raw_pp_tokens_range for the binary searches
conducted at several places in the code to work correctly. In light of
conditional inclusion directives interleaving with macro invocations, this
poses two challenges:
- First, there is no well-defined location within the sequence of pp_tokens
  originating from the macro expansion associated with such a directive.
- The directive's raw_pp_tokens_range is strictly contained within the
  raw_pp_tokens_range associated with the surrounding macro invocation. As
  all non-cond_incl_transition ops (copy, replace, insert, ...) somehow
  operating on parts or all of the sequence of pp_tokens originating from
  the macro expansion will in turn have raw_pp_tokens_ranges associated
  with them which all fully span the invocation, the cond_incl_transition
  ops corresponding to interleaving directives don't have a well-defined
  order relative to these.

In order to resolve these ambiguities:
- Give up on the assumption that all types of transformed_input_chunk::_op
  instances have a pp_tokens_range associated with them. The
  non-cond_incl_transition ops continue to do so and an associated
  raw_pp_tokens_range is determined from the pp_tokens_range by means
  of the well-defined mapping, just as before. cond_incl_transition ops
  OTOH don't have a pp_tokens_range associated with them anymore; their
  resp. raw_pp_tokens_range is to be specified at construction time and
  stored explicitly in the new _op::cond_incl_op_range_raw member. This
  change implies that it's not possible to map a pp_tokens_range to a range
  within some transformed_input_chunk's ops sequence by means of a
  binary search anymore. Drop the corresponding _find_overlapping_ops()
  overload and make its only caller, _try_rewrite_macro_arguments(),
  to search for the equivalent raw_pp_tokens_range instead.
- Let each cond_incl_transition op corresponding to some directive within a
  macro invocation assume the latter's full raw_pp_tokens_range as a first
  prerequisite step towards reestablishing a well-defined relative order
  respecting the partitioning property on the ops sequence as required for
  binary search. Note that this enforces their incomparability to any
  raw_pp_tokens_range overlapping with the macro invocation's range, which
  is indeed a strict prerequisite for maintaining the partitioning
  property: any non-cond_incl_transition ops operating on the macro
  invocation's expanded pp_tokens are incomparable to any such range
  already and as the cond_incl_transition ops will have to get inserted
  somewhere relative to these within the ops sequence, the same must hold
  for them as well.
  For establishing well-defined insertion points for cond_incl_transition
  ops within the sequence of all of a transition_chunk's ops, install the
  rule that all cond_incl_transition ops corresponding to some directive
  within a macro invocation get inserted right before the first
  non-cond_incl_transition somehow transforming that macro invocation
  expansion's resulting pp_tokens, if any. Note that the partition property
  requires that the first such non-cond_incl_transition op's associated
  raw_pp_tokens_range does not extend to before the macro invocation range.
  Make _insert_cond_incl_transition() enforce this by splitting copy ops as
  needed. Lastly, sort multiple cond_incl_transition ops all corresponding
  to directives within the same macro invocation by their directives'
  order.

Finally, the question arises how these cond_incl_transition ops
corresponding to directives within macro invocations are to be represented
in the output.
For macro invocations which can be retained in the output, the answer is
clear: the output should resemble the input 1:1. To this end, make
transformed_input_chunk::_write() look ahead and defer any
cond_incl_transition op processing to when a subsequent copy or
rewrite_macro_invocation covering said macro invocation is handled.
For macro invocations which cannot be retained in the output, the answer is
not so clear, unfortunately, because there are no predetermined points
within the sequence of (expanded) pp_tokens copied verbatim to output where
the conditional inclusion directives should be inserted. Simply write the
conditional inclusion directives directly in the order as found in the
transformed_input_chunk's ops sequence in this case, i.e. all directives
from within some macro invocation not retained in the output are emitted
right before the first (expanded) pp_token originating from that macro
invocation is being written. Note that even though most of the conditional
inclusion branches will probably turn out empty, those '#error' directives
amended by the depreprocessor to the non-taken branches for verification
purposes are still in effect.

Signed-off-by: Nicolai Stange <nstange@suse.de>