create_lp: fix issue with tag decls in struct/union/enum definitions
Whenever some declaration or definition part of a larger, surrounding
struct/union/enum definition tree is needed, the livepatch creation code
would pull in the whole such struct/union/enum definition tree for
simplicity.

However, there's a bug in the logic in that it would consider such
a struct/union/enum tree only if some dependency has been bound to one
of its descendendant tag declaration and that declaration is the first in
file order. This completely neglects the fact that a tag dependency would
get bound to the last tag declaration preceeding it in file order rather
than to the very first one. E.g. in

  struct A
  {
     struct C *c;
  };

  struct B
  {
     struct C *c;
  };

  void patched_fun(struct C* c)
  {}

the patched_fun()'s struct C parameter would get bound to the tag
declaration within struct B's definition and, as that one is not the first
one in file order, the livepatch tag dependency resolver logic would
pull in neither the definition of struct A (no dependency is bound to any
of its descendant tag declarations) nor struct B (its descendendant tag
declaration is not the first of its kind). As a result, the required
declaration of struct C at file scope would be missing alltogether from
the output.

Fix this and related cases by checking for each tag declaration dependency
whether all matching declarations preceeding it are descendants of some
larger struct/union/enum definitions and if so, emit the complete first
such struct/union/enum definition tree. Otherwise emit the first tag
declaration which is not part of a larger struct/union/enum definition
tree.

Signed-off-by: Nicolai Stange <nstange@suse.de>