preprocessor: fully expand each macro at once
Currently, the preprocessor maintains a stack of pending (recursive) macro
expansions and expands on the fly as tokens are read by the outside.

In order to simplify the state and, more importantly, to prepare for enhanced
macro expansion tracking, fully expand each macro invocation once it's being
recognized. Note that this does not yet cover the full recursive expansion yet,
but only each indivdual invocation potentially part of such a larger recursive
expansion.

As there are no more pending macro invocations from this point on, rename class
_pending_macro_expansion to _macro_expander -- this class' sole responsibility
is to implement the non-recursive expansion of a single macro invocation.

Signed-off-by: Nicolai Stange <nstange@suse.de>