create_lp: don't consider rewritten local static variables for header conflicts
When deciding about whether or not some header #include can be retained,
all non-externalized objects with static storage duration from the closure
are examined: if any of their initializers needs to get rewritten, there's a
conflict with the original declaration and the containing header is marked
as being uneligible for inclusion. However, this is overly strict in case of
static variables local to a function, because their initializer rewrites would
cause the containing function to be flagged as rewritten already. As rewritten
functions usually would get renamed as well, there's no real conflict with the
original declaration.

Ignore rewritten local static variables' initializers when determining headers'
eligibility for inclusion.

Signed-off-by: Nicolai Stange <nstange@suse.de>