ast_evaluate: allow alignof() on array types of unspecified length
Before commit 271e01f33d42 ("ast_evaluate: apply lvalue conversion from
surrounding context"), expression operands of alignof() wrongly underwent lvalue
conversions, with the effect that arrays were decayed to pointers, which are
complete types.

After this commit, array types passed to alignof() are not subject to decay
anymore, which is the desired behaviour, but makes the evaluation code to
complain about an incomplete type in case the given array is of unspecified
length.

Make the evaluation code to accept arrays of unspecified length as long as
the element type is still complete.

Signed-off-by: Nicolai Stange <nstange@suse.de>