ast_evaluate: introduce support for '__transparent_union__' attribute
The GCC __transparent_union__ attribute's usage scope is function
parameters and enables seamless passing of values to union type
function arguments without casts if the types of the former happens
to be equal to any of the union members' types.

As it's not uncommon with GNU C extension attributes, the rules of
which attributes take effect where in the AST are a bit unintuitive.
First, they can be attached to the union definition itself, at the
various places a struct/union definition can have attributes attached
to it. __transparent_union__ cannot be specified in the declaration of
function parameters. It can get specified at typedefs referencing some
union though, in which case they would apply only to the typedef.

That is,

  union __attribute__((__transparent_union__)) u
  {
    int a;
  };

  union v
  {
    int a;
  } __attribute__((__transparent_union__));

  void f(union u, union v);

  void g(void)
  {
    f(1, 1);
  }

would get accepted, as would

  union u
  {
    int a;
  };

  typedef union u ut __attribute__((__transparent_union__));

  void f(ut);

  void g(void)
  {
    f(1);
  }

That is, the __transparent_union__ feature can get enabled for the union
itself, or applied to only to "derived" type specifiers. In order to be
able to represent both case, introduce the _is_transparent_union
boolean flag to both, struct_or_union_content, which is associated
with the primary definition, if any, as well as to struct_or_union_type,
which can hold additional modifications specific to a given type spec
only.

Signed-off-by: Nicolai Stange <nstange@suse.de>