Introduce support for GCC's __auto_type extension
GCC's __auto_type extension allows for the deduction of
a declared variable's type from the initializer expression.

The constrains for __auto_type declarations are that
- there must be exactly one direct_declarator in the declaration,
- which must be of plain id declarator type
- and there must be an initializer expression of appropriate type
  present.

Different from ordinary declarations, the identifier is
not declared until after the enclosing declaration's end.
In particular, it cannot be bound to from the initializer
expression.

Introduce support for __auto_type. More specifically,
- Introduce class ast::type_specifier_auto_type, for
  representing the __auto_type specifier.
- Make the parser to recognize valid constructions involving
  __auto_type.
- Special case the resolver logic for __auto_type declarations
  to not consider the identifier declared until all after the
  containing declaration.
- Make the evaluation code to infer __auto_type variables'
  types from their associated initializer expression.

Signed-off-by: Nicolai Stange <nstange@suse.de>