Introduce support for policies implemented in Python
The current shell-out "user policy commands" based mechanism for
implementing user policies stems from the early prototyping days and has its
clear limitations:
- it's costly with respect to performance, because of the many
  forks/execve()s and
- the user policy cannot carry state between its invocations.

Introduce support for implementing livepatch creation policies in
Python. Embed a Python interpreter and implement the interfacing
code between klp-ccp and Python. Users are expected to implement the
methods defined in the LpCreationPolicyAbc Abstract Base Class (ABC)
defined in the new ccp.py. A single instance of this user specified
class will get instantiated for a klp-ccp run, and the "virtual" methods
from LpCreationPolicyAbc invoked on it as appropriate.

Users are expected to specify which class to load and use for a klp-ccp
invocation by means of the --python-policy/-P CLI option in fully
qualified form, i.e. as <MOD>.<CLASS>. In order to not require mixing of
klp-ccp Python policies with other Python packages on the file system,
especially as policies would depend on modules available only in the
embedded klp-ccp environment, there's a new environment variable
KLP_CCP_PYTHONPOLICY_PATH which allows for setting a search path for
policies.

With support for implementing livepatch creation policies in python,
rewrite the toy policy used for the testsuite to the new mechanism.

As it would have cluttered the CLI option to support both the old
and new methods of implementing the policies, and given that there's
no reason to use the old "commands" based interface at all, remove all
support for the user commands based policy mechanism.

Signed-off-by: Nicolai Stange <nstange@suse.de>