lp_creation_python_policy: supply patched funcs to python policy init
The Python policy retrieves the list of patched functions from the
environment variable KCP_FUNC. This approach is redundant, as the same
list is already passed via the command line to klp-ccp, requiring users
to specify it twice: once for klp-ccp and again for the policy via the
environment.

This patch changes the python policy API to supply the list of patched
functions directly to the Policy constructor, eliminating the need for
the policy to read it from the environment variable.

The function PyTuple_Pack internally increments the reference to
'patched_funcs_py_list', so decrement its reference right after
assigning it to the tuple.

Signed-off-by: Vincenzo MEZZELA <vincenzo.mezzela@suse.com>