examples: ipa-clones: skip "(null)" entry in ipa-clones dump
As for [1], the output of -fdump-ipa-clones can contain "(null)" entries
that shouldn't be there (as mentioned in the bug). Assuming [1] is the
only source of all the "(null)" entry, it should be safe to just skip
these.

Hitting such an entry during the ipa-clones analysis in the policies
would cause them to raise an error, leading klp-ccp to exit.

- [1] https://gcc.gnu.org/bugzilla/show_bug.cgi?id=119852

Signed-off-by: Vincenzo Mezzela <vincenzo.mezzela@suse.com>