create_lp: Simplify header inclusion heuristic
The current header inclusion heuristics always adds an unnecesary
prologue of low-level macros to the generated livepatch files,
which have to be manually removed every time to successfully
compile the code.

As a way to fix the issue, simplify the logic and retain all
includes from the input source file.

Signed-off-by: Fernando Gonzalez <fernando.gonzalez@suse.com>